requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
PyJWT==2.8.0
python-dotenv==1.0.1
google-genai
deepgram-sdk
pypdf
python-docx
python-multipart

# Ported from Rahat
groq
pdfplumber
python-docx
beautifulsoup4
pillow
pytesseract

